api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
kafka-python==2.0.2
pymongo==4.8.0
requests==2.32.3
